Exit Abfrage mit Endlosschleifenbug und grober None Behebung.

+++ test.py
@@ -19,13 +19,23 @@ def name():
 def close():
   password = input("Your password: ")
   while password != "Python":
-    password = input("Your Password: ")
-  os.kill(os.getpid(), signal.SIGTERM) 
-print(test())
+    password = input("Your Password: ") 
+  return " "
+test()
 print(name())
 print("Last step...")
 print("Feierabend")
 print("Test vom Gemini")
-print("Hello from the Beans Corporation Terminaltron CE!!!")
+print("Hello from the Beans Corporation Terminaltron CE!!!") 
 print(close())
+sdownchoice = input("Exit (w)ith or with(o)ut delay???")
+while sdownchoice != "w" or "o":
+  if sdownchoice == "w":
+     time.sleep(5)
+     os.kill(os.getpid(), signal.SIGTERM)
+  if sdownchoice == "o":
+     os.kill(os.getpid(), signal.SIGTERM)
+  else:
+     print("Don't try shit with me!!!")
+
